move repository method to orm adapter
